Revert JDK-8268119: Rename copy_os_cpu.inline.hpp files to copy_os_cpu.hpp && JDK-8142362: Lots of code duplication in Copy class
